Extract file's operation to FileProvider.kt
